refractor test file structure. Changes to be committed: 	modified:   database_setup.py 	new file:   database_tear.py 	new file:   tests/__init__.py 	modified:   tests/test_endpoints.py 	modified:   tests/test_validations.py
